#106: extract IotsaWifiDriver, restoring IotsaWifiMod to standard-module concerns
Mechanical split, no behaviour change (all builds green: esp8266/esp32/c3/
no-networking/default).

- New IotsaWifiDriver (src/iotsaWifiDriver.{h,cpp}) -- a plain object (not an
  IotsaBaseModule) holding the mechanism: startStation/stopStation/startAP/
  stopAP/reinitStack, readActualState(), drainEvents(), begin() (installs the
  platform WiFi event handlers), and the volatile latch storage. The shared
  value types (IotsaWifiActualState/Events/StaFailReason) move here from
  iotsaWifiController.h. Its one input from the module is setTxPowerReduction().
- IotsaWifiController::_mod -> _driver (IotsaWifiDriver&); ctor now takes the
  driver.
- IotsaWifiMod drops the driver-surface public API and all the latch members;
  it owns `IotsaWifiDriver _driver` + `IotsaWifiController _controller{_driver}`
  and keeps only config load/save, the /api/wificonfig + web interface, info(),
  and lifecycle. setup() calls _driver.begin(); configLoad()/configSave() push
  wifiPowerReduction into the driver.

So the three concerns are cleanly separated: driver = platform/radio,
controller = policy, module = iotsa integration + persisted settings.
